Split gui/journal into logical modules
